+++ MD_steps/s5_other/control.c
@@ -46,8 +46,8 @@ int main(int argc, char *argv[]) {
   for (i = 0; i < Nbody; i++) {
     fscanf(in, "%16le%16le%16le%16le%16le%16le%16le%16le%16le\n", mass + i,
            radius + i, vis + i, &pos[i][Xcoord], &pos[i][Ycoord],
-           &pos[i][Zcoord], &velo[Xcoord][i], &velo[Ycoord][i],
-           &velo[Zcoord][i]);
+           &pos[i][Zcoord], &velo[i][Xcoord], &velo[i][Ycoord],
+           &velo[i][Zcoord]);
   }
   fclose(in);
 
@@ -75,8 +75,8 @@ int main(int argc, char *argv[]) {
     for (i = 0; i < Nbody; i++) {
       fprintf(out, "%16.8E%16.8E%16.8E%16.8E%16.8E%16.8E%16.8E%16.8E%16.8E\n",
               mass[i], radius[i], vis[i], pos[i][Xcoord], pos[i][Ycoord],
-              pos[i][Zcoord], velo[Xcoord][i], velo[Ycoord][i],
-              velo[Zcoord][i]);
+              pos[i][Zcoord], velo[i][Xcoord], velo[i][Ycoord],
+              velo[i][Zcoord]);
     }
     fclose(out);
   }